fix_enhancement_119: Вынести обработку событий в отдельную функцию, для уменьшения вложенности и повышения удобочитаемость кода.
